+++ src/core/grid_based_algorithms/lbgpu.cuh
@@ -12,7 +12,7 @@
 typedef struct {
 
   /** velocity density of the node */
-  float *vd;
+  double *vd;
   /** seed for the random gen */
   unsigned int *seed;
   /** flag indicating whether this site belongs to a boundary */
